requirements.txt
torch
transformers
datasets
faiss-cpu
langchain
sentence-transformers
pypdf